refactor: move SubjectCardList to ArchivedComponents
