+++ src/components/ArchivedComponents/SubjectCardList/SubjectCardListItem.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
+import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
 
 import { KeyStageSubject } from "./SubjectCardList";
 import SubjectCardListItem from "./SubjectCardListItem";
