+++ src/components/ArchivedComponents/SubjectCardList/SubjectCardListItem.tsx
@@ -1,11 +1,11 @@
 import { FC } from "react";
 
-import useAnalytics from "../../context/Analytics/useAnalytics";
-import type { KeyStageTitleValueType } from "../../browser-lib/avo/Avo";
-import useClickableCard from "../../hooks/useClickableCard";
-import OakLink from "../OakLink";
-import SubjectIcon from "../SharedComponents/SubjectIcon";
-import useAnalyticsPageProps from "../../hooks/useAnalyticsPageProps";
+import useAnalytics from "../../../context/Analytics/useAnalytics";
+import type { KeyStageTitleValueType } from "../../../browser-lib/avo/Avo";
+import useClickableCard from "../../../hooks/useClickableCard";
+import OakLink from "../../OakLink";
+import SubjectIcon from "../../SharedComponents/SubjectIcon";
+import useAnalyticsPageProps from "../../../hooks/useAnalyticsPageProps";
 
 import { KeyStageSubject } from "./SubjectCardList";
